Safari-611.1.10/LayoutTests/fast/css/stylesheet-parameters-expected.txt

Tests that the parameters to CSSStyleSheet.insertRule() / deleteRule() are mandatory.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS stylesheet.__proto__ is CSSStyleSheet.prototype
PASS stylesheet.deleteRule() threw exception TypeError: Not enough arguments.
PASS stylesheet.insertRule() threw exception TypeError: Not enough arguments.
PASS stylesheet.rules.length is 0
PASS stylesheet.insertRule('body { margin: 0; }') did not throw exception.
PASS stylesheet.rules.length is 1
PASS successfullyParsed is true

TEST COMPLETE
